+++ packages/ai/core/generate-text/output.ts
@@ -1,11 +1,14 @@
-import { parseJSON } from '@ai-sdk/provider-utils';
+import { safeParseJSON, safeValidateTypes } from '@ai-sdk/provider-utils';
 import { asSchema, Schema } from '@ai-sdk/ui-utils';
 import { z } from 'zod';
+import { NoObjectGeneratedError } from '../../errors';
 import { injectJsonInstruction } from '../generate-object/inject-json-instruction';
 import {
   LanguageModel,
   LanguageModelV1CallOptions,
 } from '../types/language-model';
+import { LanguageModelResponseMetadata } from '../types/language-model-response-metadata';
+import { LanguageModelUsage } from '../types/usage';
 
 export interface Output<OUTPUT> {
   readonly type: 'object' | 'text';
@@ -16,7 +19,13 @@ export interface Output<OUTPUT> {
   responseFormat: (options: {
     model: LanguageModel;
   }) => LanguageModelV1CallOptions['responseFormat'];
-  parseOutput(options: { text: string }): OUTPUT;
+  parseOutput(
+    options: { text: string },
+    context: {
+      response: LanguageModelResponseMetadata;
+      usage: LanguageModelUsage;
+    },
+  ): OUTPUT;
 }
 
 export const text = (): Output<string> => ({
@@ -53,8 +62,41 @@ export const object = <OUTPUT>({
             schema: schema.jsonSchema,
           });
     },
-    parseOutput({ text }: { text: string }) {
-      return parseJSON({ text, schema });
+    parseOutput(
+      { text }: { text: string },
+      context: {
+        response: LanguageModelResponseMetadata;
+        usage: LanguageModelUsage;
+      },
+    ) {
+      const parseResult = safeParseJSON({ text });
+
+      if (!parseResult.success) {
+        throw new NoObjectGeneratedError({
+          message: 'No object generated: could not parse the response.',
+          cause: parseResult.error,
+          text,
+          response: context.response,
+          usage: context.usage,
+        });
+      }
+
+      const validationResult = safeValidateTypes({
+        value: parseResult.value,
+        schema,
+      });
+
+      if (!validationResult.success) {
+        throw new NoObjectGeneratedError({
+          message: 'No object generated: response did not match schema.',
+          cause: validationResult.error,
+          text,
+          response: context.response,
+          usage: context.usage,
+        });
+      }
+
+      return validationResult.value;
     },
   };
 };