+++ packages/ai/core/prompt/convert-to-language-model-prompt.ts
@@ -284,20 +284,23 @@ function convertPartToLanguageModelPart(
   // Now that we have the normalized data either as a URL or a Uint8Array,
   // we can create the LanguageModelV1Part.
   switch (type) {
-    case 'image':
-      // We give a best effort to detect the mime type if it is not provided.
-      // otherwise, we use the provided mime type.
-      if (mimeType == null && normalizedData instanceof Uint8Array) {
-        mimeType = detectImageMimeType(normalizedData);
-      }
+    case 'image': {
+      // When possible, try to detect the mimetype automatically
+      // to deal with incorrect mimetype inputs.
+      // When detection fails, use provided mimetype.
 
+      if (normalizedData instanceof Uint8Array) {
+        mimeType = detectImageMimeType(normalizedData) ?? mimeType;
+      }
       return {
         type: 'image',
         image: normalizedData,
         mimeType,
         providerMetadata: part.experimental_providerMetadata,
       };
-    case 'file':
+    }
+
+    case 'file': {
       // We should have a mimeType at this point, if not, throw an error.
       if (mimeType == null) {
         throw new Error(`Mime type is missing for file part`);
@@ -312,5 +315,6 @@ function convertPartToLanguageModelPart(
         mimeType,
         providerMetadata: part.experimental_providerMetadata,
       };
+    }
   }
 }