+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,54 @@
 # ai
 
+## 4.0.23
+
+### Patch Changes
+
+- bc4cd19: feat (ai/core): consolidate whitespace in smooth stream
+
+## 4.0.22
+
+### Patch Changes
+
+- Updated dependencies [5ed5e45]
+  - @ai-sdk/provider-utils@2.0.5
+  - @ai-sdk/provider@1.0.3
+  - @ai-sdk/react@1.0.7
+  - @ai-sdk/ui-utils@1.0.6
+
+## 4.0.21
+
+### Patch Changes
+
+- a8669a2: fix (ai/core): prefer auto-detected image mimetype
+- 6fb3e91: fix (ai/core): include type in generateText toolResults result property.
+
+## 4.0.20
+
+### Patch Changes
+
+- da9d240: fix (ai/core): suppress errors caused by writing to closed stream
+- 6f1bfde: fix (ai/core): invoke streamText tool call repair when tool cannot be found
+
+## 4.0.19
+
+### Patch Changes
+
+- c3a6065: fix (ai/core): apply transform before callbacks and resolvables
+
+## 4.0.18
+
+### Patch Changes
+
+- 304e6d3: feat (ai/core): standardize generateObject, streamObject, and output errors to NoObjectGeneratedError
+- 304e6d3: feat (ai/core): add additional information to NoObjectGeneratedError
+
+## 4.0.17
+
+### Patch Changes
+
+- 54bbf21: fix (ai/core): change streamText.experimental_transform signature to support tool type inference
+
 ## 4.0.16
 
 ### Patch Changes