+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,57 @@
 
 ### Patch Changes
 
+- Updated dependencies [bc4cd19]
+  - ai@4.0.23
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.0.22
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [a8669a2]
+- Updated dependencies [6fb3e91]
+  - ai@4.0.21
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [da9d240]
+- Updated dependencies [6f1bfde]
+  - ai@4.0.20
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [c3a6065]
+  - ai@4.0.19
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [304e6d3]
+- Updated dependencies [304e6d3]
+  - ai@4.0.18
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [54bbf21]
+  - ai@4.0.17
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [e3fac3f]
   - ai@4.0.16