+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @ai-sdk/google-vertex
 
+## 2.0.12
+
+### Patch Changes
+
+- 5ed5e45: chore (config): Use ts-library.json tsconfig for no-UI libs.
+- Updated dependencies [5ed5e45]
+  - @ai-sdk/provider-utils@2.0.5
+  - @ai-sdk/anthropic@1.0.6
+  - @ai-sdk/provider@1.0.3
+  - @ai-sdk/google@1.0.12
+
+## 2.0.11
+
+### Patch Changes
+
+- 5feec50: feat (provider/google-vertex): Add imagen support.
+
+## 2.0.10
+
+### Patch Changes
+
+- d32abbd: feat (provider/google-vertex): Add gemini 2 models.
+
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [db31e74]
+  - @ai-sdk/google@1.0.11
+
 ## 2.0.8
 
 ### Patch Changes