+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/provider-utils
 
+## 2.0.5
+
+### Patch Changes
+
+- 5ed5e45: chore (config): Use ts-library.json tsconfig for no-UI libs.
+- Updated dependencies [5ed5e45]
+  - @ai-sdk/provider@1.0.3
+
 ## 2.0.4
 
 ### Patch Changes